--- src/app/helpers/use-page-data.d.ts
@@ -1,12 +1,0 @@
-import type {LocaleEntry} from '~/components/language-selector/language-selector';
-
-type PageData = {
-    error?: string;
-    translations: [Array<LocaleEntry>];
-};
-
-export default function usePageData(
-    slug: string,
-    preserveWrapping?: boolean,
-    noCamelCase?: boolean
-): PageData;
